streamlit/requirements.txt
# dependencies
streamlit
snowflake-snowpark-python
snowflake-snowpark-python[pandas]